Move SigningKey into sol.rs and enable outside of cfg(test)
