+++ state-chain/chains/src/sol/instruction_builder.rs
@@ -364,11 +364,10 @@ mod test {
 	use crate::{
 		sol::{
 			sol_tx_core::{
-				address_derivation::derive_deposit_address,
-				extra_types_for_testing::{Signer, SigningKey},
+				address_derivation::derive_deposit_address, extra_types_for_testing::Signer,
 				sol_test_values::*,
 			},
-			SolHash, SolMessage, SolTransaction, SolanaDepositFetchId,
+			SigningKey, SolHash, SolMessage, SolTransaction, SolanaDepositFetchId,
 		},
 		TransferAssetParams,
 	};
